requirements.txt
semantic-kernel>=0.9.0b1
langchain>=0.1.0
langchain-openai>=0.0.3
langchain-community>=0.0.6
langchain-experimental>=0.0.26
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pyodbc>=4.0.39
